Make appending instructions to blocks more efficient
Previously, this ran in O(n^2) (where n is the number of instructions
to be added) as new instructions were inserted at the end of the list
representing the instruction stack. As this list grows longer, appends
become more expensive, involving a bunch of copying and GC overhead.
This change switches to prepending instructions to the stack (with a
reverse call before converting to BasicBlock), which is an O(1)
operation, yielding O(n) for our n instructions.